# Small performance optimizations for formatting stats